+++ src/wavefilt.h
@@ -17,12 +17,12 @@ extern "C" {
 
 int filtlength(const char* name);
 
-int filtcoef(const char* name, double *lp1, double *hp1, double *lp2, double *hp2);
+int filtcoef(const char* name, float *lp1, float *hp1, float *lp2, float *hp2);
 
-void copy_reverse(const double *in, int N, double *out);
-void qmf_even(const double *in, int N, double *out);
-void qmf_wrev(const double *in, int N, double *out);
-void copy(const double *in, int N, double *out);
+void copy_reverse(const float *in, int N, float *out);
+void qmf_even(const float *in, int N, float *out);
+void qmf_wrev(const float *in, int N, float *out);
+void copy(const float *in, int N, float *out);
 #ifdef __cplusplus
 }
 #endif
